Fixed /startkick Fixed /mutep error in console Console availability added to /startkick and /gamemode
Took 1 hour 1 minute

+++ src/main/java/de/crashmash/citybuild/storage/Storage.java
@@ -47,7 +47,7 @@ public class Storage {
         this.startKickCollection = database.createCollection("StartKick")
                 .field("UUID", DataType.UUID, FieldOption.PRIMARY_KEY)
                 .field("Reason", DataType.STRING)
-                .field("Duration", DataType.INTEGER)
+                .field("Duration", DataType.LONG)
                 .field("Cooldown", DataType.LONG)
                 .create();
